Ajuste login e atendimento para funcionamento com mariaDB * utilizar o arquivo /sql/criacaoTabelas.sql e trocar senha do root para funcionar

+++ src/Controller/TelaConsultaController.java
@@ -58,7 +58,7 @@ public class TelaConsultaController{
 	public Consulta toEntity() {
 		Consulta consulta = new Consulta();
 		
-		consulta.setId(id.get());
+		consulta.setId((int) id.get());
 		consulta.setCpf(cpf.get());
 		consulta.setCrm(crm.get());
 		consulta.setData((LocalDate) dataConsult.get());
